lexer: create external token_clear function
